requirements.txt
altgraph==0.17.2
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==35.0.0
cycler==0.10.0
decorator==4.4.2
docutils==0.16
idna==2.10
image-slicer==2.1.1
imageio==2.9.0
importlib-metadata==4.8.2
jeepney==0.7.1
keyring==23.2.1
kiwisolver==1.3.2
matplot==0.1.9
matplotlib==3.3.3
mpmath==1.1.0
networkx==2.5
numpy==1.19.5
opencv-python==4.5.1.48
packaging==20.8
Pillow==7.2.0
pkg_resources==0.0.0
pkginfo==1.6.1
pycparser==2.21
Pygments==2.7.3
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.3
pyloco==0.0.139
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
readme-renderer==28.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
scikit-image==0.18.1
scipy==1.6.0
SecretStorage==3.3.1
SimpleWebSocketServer==0.1.1
six==1.15.0
sympy==1.7.1
tifffile==2021.1.14
torch==1.10.0+cpu
torchvision==0.11.1+cpu
tqdm==4.55.1
twine==3.6.0
urllib3==1.26.7
ushlex==0.99.1
webencodings==0.5.1
websocket-client==1.2.1
zipp==3.6.0
